Revert "memory: add depth info to node and decrease heapid size to uint16"
This reverts commit 829f073218aef49b6550b5731dcb80d00b628f89.

+++ src/ak_memory_lt.h
@@ -14,7 +14,7 @@
 
 typedef struct AkHeapBucketEntry {
   AkHeap  *heap;
-  uint16_t heapid;
+  uint32_t heapid;
 } AkHeapBucketEntry;
 
 typedef struct AkHeapBucket {
@@ -38,10 +38,10 @@ void
 ak_heap_lt_init(AkHeap * __restrict initialHeap);
 
 AkHeap *
-ak_heap_lt_find(uint16_t heapid);
+ak_heap_lt_find(uint32_t heapid);
 
 void
-ak_heap_lt_remove(uint16_t heapid);
+ak_heap_lt_remove(uint32_t heapid);
 
 void
 ak_heap_lt_insert(AkHeap * __restrict heap);
